+++ lib/commands/server_command.rb
@@ -19,13 +19,28 @@ module Staticz
       desc "Print this page"
     end
 
+    option :port do
+      desc "Define the port used for the webserver"
+      short "-p"
+      long "--port port"
+
+      default 3000
+      convert :integer
+    end
+
     def run
       if params[:help]
         print help
-      else
-        Staticz::Settings.development!
-        Staticz::Server.new
+        exit
       end
+
+      if !params[:port]
+        puts "A given port must be an integer"
+        exit 1
+      end
+
+      Staticz::Settings.development!
+      Staticz::Server.new(params[:port])
     end
   end
 end